requirements.txt
numpy
tensorflow
tensorflow-datasets
pyyaml
Pillow
